+++ src/constants.rs
@@ -24,7 +24,7 @@ pub static LABELS: [&str; 10] = [
     "Blue juice",
     "Special Leaps",
 ];
-pub static NORMAL_HONE_CHANCES: [f32; 25] = [
+pub static NORMAL_HONE_CHANCES: [f64; 25] = [
     1.0, 1.0, 1.0, 0.45, 0.45, 0.45, 0.3, 0.3, 0.15, 0.15, 0.1, 0.1, 0.05, 0.05, 0.04, 0.04, 0.03,
     0.03, 0.03, 0.015, 0.015, 0.01, 0.01, 0.005, 0.005,
 ];
